♻️ [Refactor]: feature/landing 제거 및 fsd 방식으로 리팩토링 (빌드오류제거)
