+++ src/app/(auth)/study-room/[id]/manage-applicants/page.tsx
@@ -1,4 +1,4 @@
-import ManageCard from "@/app/(auth)/study-room/components/manage-applicants/ManageCard";
+import ManageCard from "@/entities/study/join/ui/ManageCard";
 
 const ManageApplicants = () => {
   return <ManageCard />;
