+++ src/features/study/studies/ui/studies.ui.types.ts
@@ -1,0 +1,6 @@
+import { Study } from "@/entities/study/studies/api/studies.api.type";
+import { HTMLAttributes } from "react";
+
+export interface StudyCardProps extends HTMLAttributes<HTMLDivElement> {
+  data: Study;
+}
