+++ src/features/study/study-filter/lib/index.ts
@@ -1,0 +1,12 @@
+export const meetTypes = [
+  { label: "전체", value: null },
+  { label: "대면", value: "대면" },
+  { label: "비대면", value: "비대면" },
+  { label: "혼합", value: "혼합" },
+];
+
+export const recruitmentTypes = [
+  { label: "전체", value: null },
+  { label: "모집 중", value: "모집 중" },
+  { label: "모집 완료", value: "모집 완료" },
+];
