+++ src/widget/landing/ui/landing.ui.types.ts
@@ -1,0 +1,5 @@
+import { GetStudyCategoriesResponse } from "@/entities/study/category/api/category.api.types";
+
+export interface LandingContentHeaderProps {
+  categories: GetStudyCategoriesResponse;
+}
